refactor(drawingml): extract ShapePropertiesWriter from the chart series (spec 16 task 3)
A chart series' c:spPr and a shape's xdr:sp/spPr are the same schema type,
a:CT_ShapeProperties. The rules for writing into it are subtle and were
implemented once, correctly, inside the chart series module. Spec 15 needs them
for shapes, and a second implementation is how the two copies drift.

ShapePropertiesWriter now owns them, typed over OpenXmlCompositeElement so no
host's element type appears in its signatures:

    SetFill(shapeProperties, XLColor?)
    EnsureOutline(shapeProperties) -> A.Outline
    SetOutlineWidth(outline, double?)
    SetOutlineColor(outline, XLColor?)

with BuildColor, MapSchemeColor and the CT_ShapeProperties sequence predicates
private to it. Consumers get theme-colour writing for free, because the colour
building moved with the setters.

This is not a mechanical move. SetOutline used to take the series and its flags
enum and decide for itself which of the width and the colour to write.
The mask now stops at the chart boundary: ChartSeriesFormatXml keeps a thin
adapter that reads XLChartSeriesFormat and asks for the operations it names, and
the shared layer is told what to set rather than what was assigned. No flags enum
appears in a DrawingML signature, and nothing in DrawingML/ mentions a chart type
at all.

Parent positioning stayed with the parent. Where c:spPr sits inside c:ser is
chart knowledge, so NewShapeProperties and its type list are untouched; the
shared layer only ever operates within an spPr it is handed.

DrawingUnits gains EmuPerPoint and PointsToEmu, so the line-width conversion has
one definition rather than one on each side of the new boundary.

Two of the items the spec listed for this task did not move, and the reasons are
findings rather than omissions -- see the report accompanying this commit:

- AppendBeforeExtensionList names C.ExtensionList, a chart-namespace type.
  Moving it into DrawingML/ would have put a chart type into a shared signature,
  the same category of mistake as putting a flags enum there.
- InsertAfterLastOf could have moved, but spec 22 -- which post-dates the writing
  of spec 16 -- already introduced ChartElementOrder.InsertOrdered, a better
  generalisation of both helpers that the sibling chart concept modules already
  use. Moving the older helper into DrawingML/ would have made three ordering
  implementations where there are currently two. Consolidating the two chart-side
  is the right follow-up and belongs to spec 22's territory, not here.

Behaviour preserved:

- The nine patch-* goldens are byte-identical: a fill replaced, cleared and
  themed; an outline widened, recoloured and added; a gradient replaced through
  the choice group; c:spPr built from nothing; the marker's own shape properties.
- All 148 chart tests green and unmodified. No test was added, removed or changed
  by this commit.
- 29146 tests green across the solution on net8.0 and net10.0, unchanged in count
  from before the task.

Satisfies acceptance criteria 1 (ShapePropertiesWriter is internal;
PublicAPI.Unshipped.txt untouched), 2, 3 (golden byte-identity), 4 (no flags enum
in a DrawingML signature) and 6 (no operation added; the layer's observable
capability set is what the chart suite already exercised).
